Move markdown file to correct package
